Move Reflection into QuoteContext
* Make Reflection the module class of QuoteContext.reflect
