+++ tests/neg-macros/i6324.scala
@@ -1,5 +1,5 @@
 class Test {
-  def res(x: quoted.Expr[Int])(using quoted.Reflection): quoted.Expr[Int] = x match {
+  def res(x: quoted.Expr[Int])(using quoted.Reflection): quoted.Expr[Int] = x match { // error
     case '{ 1 + $b } => // error: Type must be fully defined. Consider annotating the splice using a type ascription: (${b}: XYZ).
       b // error: Not found: b
   }
