+++ tests/neg-macros/i6325.scala
@@ -1,5 +1,5 @@
 object Test {
-  def res(x: quoted.Expr[Int])(using quoted.Reflection): quoted.Expr[Int] = x match {
+  def res(x: quoted.Expr[Int])(using quoted.Reflection): quoted.Expr[Int] = x match { // error
     case '{ 1 + (${Bind(b)}: Int) } => ??? // error: Not found: Bind
     case _ => ???
   }
