+++ tests/pos-macros/i9240/Macro_1.scala
@@ -4,10 +4,10 @@ inline def diveInto[T]: String = ${ diveIntoImpl[T]() }
 
 def diveIntoImpl[T]()(implicit qctx: QuoteContext, ttype: Type[T]): Expr[String] =
   import qctx.reflect._
-  Expr( unwindType(qctx.reflect)(TypeRepr.of[T]) )
+  Expr( unwindType(TypeRepr.of[T]) )
 
-def unwindType(reflect: Reflection)(aType: reflect.TypeRepr): String =
-  import reflect._
+def unwindType(using QuoteContext)(aType: qctx.reflect.TypeRepr): String =
+  import qctx.reflect._
 
   aType match {
     case AppliedType(t,tob) =>
